+++ multi/init.lua
@@ -28,7 +28,7 @@ multi.Version = "13.0.0"
 multi._VERSION = "13.0.0"
 multi.stage = "stable"
 multi.__index = multi
-multi.Name = "multi.Root"
+multi.Name = "multi.root"
 multi.Mainloop = {}
 multi.Garbage = {}
 multi.ender = {}
@@ -121,7 +121,7 @@ function multi:enableLoadDetection()
 	local t = os.clock()
 	local stop = false
 	temp:benchMark(.01):OnBench(function(time,steps)
-		stop = steps*1.1
+		stop = steps
 	end)
 	while not stop do
 		temp:uManager()
@@ -135,6 +135,7 @@ function multi:setLoad(n)
 end
 local busy = false
 local lastVal = 0
+local bb = 0
 function multi:getLoad()
 	if not multi.maxSpd then multi:enableLoadDetection() end
 	if busy then return lastVal end
@@ -143,6 +144,7 @@ function multi:getLoad()
 		local bench
 		multi:benchMark(.01):OnBench(function(time,steps)
 			bench = steps
+			bb = steps
 		end)
 		thread.hold(function()
 			return bench
@@ -154,6 +156,7 @@ function multi:getLoad()
 		local bench
 		multi:benchMark(.01):OnBench(function(time,steps)
 			bench = steps
+			bb = steps
 		end)
 		while not bench do 
 			multi:uManager()
@@ -165,7 +168,7 @@ function multi:getLoad()
 	if val<0 then val = 0 end
 	if val > 100 then val = 100 end
 	lastVal = val
-	return val
+	return val,bb*100
 end
 function multi:setDomainName(name)
 	self[name]={}
@@ -197,6 +200,13 @@ function multi:setPriority(s)
 		end
 		self.solid = true
 	end
+	if not self.PrioritySet then
+		self.defPriority = self.Priority
+		self.PrioritySet = true
+	end
+end
+function multi:ResetPriority()
+	self.Priority = self.defPriority
 end
 -- System
 function os.getOS()
@@ -345,20 +355,24 @@ function multi:getTasksDetails(t)
 				dat2 = dat2.."<SystemThread: "..multi.SystemThreads[i].Name.." | "..os.clock()-multi.SystemThreads[i].creationTime..">\n"
 			end
 		end
+		local load,steps = multi:getLoad()
 		if multi.scheduler then
 			for i=1,#multi.scheduler.Threads do
 				dat = dat .. "<THREAD: "..multi.scheduler.Threads[i].Name.." | "..os.clock()-multi.scheduler.Threads[i].creationTime..">\n"
 			end
-			return "Load on "..ProcessName[self.Type=="process"].."<"..(self.Name or "Unnamed")..">"..": "..multi.Round(multi:getLoad(),2).."%\nMemory Usage: "..math.ceil(collectgarbage("count")).." KB\nThreads Running: "..#multi.scheduler.Threads.."\nSystemThreads Running: "..#(multi.SystemThreads or {}).."\nPriority Scheme: "..priorityTable[multi.defaultSettings.priority or 0].."\n\n"..s.."\n\n"..dat..dat2
+			return "Load on "..ProcessName[self.Type=="process"].."<"..(self.Name or "Unnamed")..">"..": "..multi.Round(load,2).."%\nCycles Per Second Per Task: "..steps.."\nMemory Usage: "..math.ceil(collectgarbage("count")).." KB\nThreads Running: "..#multi.scheduler.Threads.."\nSystemThreads Running: "..#(multi.SystemThreads or {}).."\nPriority Scheme: "..priorityTable[multi.defaultSettings.priority or 0].."\n\n"..s.."\n\n"..dat..dat2
 		else
-			return "Load on "..ProcessName[self.Type=="process"].."<"..(self.Name or "Unnamed")..">"..": "..multi.Round(multi:getLoad(),2).."%\nMemory Usage: "..math.ceil(collectgarbage("count")).." KB\nThreads Running: 0\nPriority Scheme: "..priorityTable[multi.defaultSettings.priority or 0].."\n\n"..s..dat2
+			return "Load on "..ProcessName[self.Type=="process"].."<"..(self.Name or "Unnamed")..">"..": "..multi.Round(load,2).."%\nCycles Per Second Per Task: "..steps.."\n\nMemory Usage: "..math.ceil(collectgarbage("count")).." KB\nThreads Running: 0\nPriority Scheme: "..priorityTable[multi.defaultSettings.priority or 0].."\n\n"..s..dat2
 		end
 	elseif t == "t" or t == "table" then
+		local load,steps = multi:getLoad()
 		str = {
+			ProcessName = (self.Name or "Unnamed"),
 			ThreadCount = #multi.scheduler.Threads,
 			MemoryUsage = math.ceil(collectgarbage("count")).." KB",
 			PriorityScheme = priorityTable[multi.defaultSettings.priority or 0],
-			SystemLoad = multi.Round(multi:getLoad(),2)
+			SystemLoad = multi.Round(load,2),
+			CyclesPerSecondPerTask = steps,
 		}
 		str.threads = {}
 		str.systemthreads = {}
@@ -970,7 +984,7 @@ end
 function multi:newAlarm(set)
 	local c=self:newBase()
 	c.Type='alarm'
-	c.Priority=self.Priority_Low
+	c:setPriority("Low")
 	c.set=set or 0
 	local count = 0
 	local t = clock()
@@ -1174,7 +1188,7 @@ function multi:newTStep(start,reset,count,set)
 	local c=self:newBase()
 	think=1
 	c.Type='tstep'
-	c.Priority=self.Priority_Low
+	c:setPriority("Low")
 	c.start=start or 1
 	local reset = reset or math.huge
 	c.endAt=reset
@@ -1266,7 +1280,7 @@ function multi:newTimeStamper()
 		["12"] = 31,
 	}
 	c.Type='timestamper'
-	c.Priority=self.Priority_Idle
+	c:setPriority("Idle")
 	c.hour = {}
 	c.minute = {}
 	c.second = {}
@@ -1886,8 +1900,8 @@ function multi:mainloop(settings)
 		local PS=self
 		local PStep = 1
 		local autoP = 0
-		local solid
-		local sRef
+		local solid,sRef
+		local cc=0
 		while mainloopActive do
 			if next then
 				local DD = table.remove(next,1)
@@ -1947,8 +1961,12 @@ function multi:mainloop(settings)
 					PStep=0
 				end
 			elseif priority == 3 then
-				tt = clock()-t
-				t = clock()
+				cc=cc+1
+				if cc == 1000 then
+					tt = clock()-t
+					t = clock()
+					cc=0
+				end
 				for _D=#Loop,1,-1 do
 					if Loop[_D] then
 						if Loop[_D].Priority == p_c or (Loop[_D].Priority == p_h and tt<.5) or (Loop[_D].Priority == p_an and tt<.125) or (Loop[_D].Priority == p_n and tt<.063) or (Loop[_D].Priority == p_bn and tt<.016) or (Loop[_D].Priority == p_l and tt<.003) or (Loop[_D].Priority == p_i and tt<.001) then